+++ components/Carousel/Carousel.tsx
@@ -1,19 +1,13 @@
 import { Carousel as Slider } from '@trendyol-js/react-carousel';
 import Image from "next/image";
-// images testing data:
-const imgRes = [
-    "https://placeimg.com/300/300/any",
-    "https://placeimg.com/300/300/animals",
-    "https://placeimg.com/300/300/architecture",
-    "https://placeimg.com/300/300/nature",
-    "https://placeimg.com/300/300/people",
-    "https://placeimg.com/300/300/tech"
-]
 
-const Carousel = () => {
+interface CaroProps{
+    imagesArray: string[]
+}
+const Carousel = ({imagesArray}:CaroProps) => {
     return (
         <Slider show={1} slide={1} swiping={true}  >
-            {imgRes.map((img, index) => {
+            {imagesArray.map((img, index) => {
                 return (
                     <Image
                         key={index}
