+++ pages/index.tsx
@@ -2,10 +2,10 @@ import type { NextPage } from "next";
 import Head from "next/head";
 import Image from "next/image";
 import styles from "../styles/Home.module.css";
-import Item from "./item";
+import ProductPage from "./productPage";
 
 const Home: NextPage = () => {
-  return <Item />;
+  return <ProductPage />;
 };
 
 export default Home;
